invent_pc/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
requests==2.26.0
python-dotenv==1.0.0
psycopg2-binary==2.9.3
python-dotenv==1.0.0